+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Support opaque types in doctest results (e.g. `#User<name: "", ...>`).
+
 ## 0.2.1 (2024-03-22)
 
 - Do not crash if doctest has no expected result.